fix: errors in authentication

+++ src/lib/Auth.js
@@ -6,12 +6,18 @@ export async function request(path, data = {}){
         headers: {"Content-Type": "application/json"},
         body: JSON.stringify(data)
     })
+
+    let response = await res.json();
+
     if(!res.ok){
+        if(response.error !== 'Authentication'){
+            return {error: response.error}
+        }
         localStorage.clear()
         window.location.reload()
-        return {}
+        return {error: response.error}
     }
 
-    return await res.json()
+    return response
 }
